# Conflicts:
#   BACK/sinbangsa/src/main/java/com/sinbangsa/data/repository/StoreRepository.java
#   BACK/sinbangsa/src/main/java/com/sinbangsa/data/repository/ThemeRepository.java
#   BACK/sinbangsa/src/main/java/com/sinbangsa/data/repository/ThemeReviewRepository.java
#   BACK/sinbangsa/src/main/java/com/sinbangsa/data/repository/UserStoreRelationRepository.java

+++ BACK/sinbangsa/src/main/java/com/sinbangsa/data/entity/Book.java
@@ -30,7 +30,7 @@ public class Book {
 
     @Column
     @NotNull
-    private long clear;
+    private int clear;
 
     @Column
     @NotNull
